User-Task relation

+++ src/server/router/taskRouter.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { createProtectedRouter } from "./protected-router";
+import { createProtectedRouter } from './protected-router';
 
 export const taskRouter = createProtectedRouter()
   // CREATE
@@ -7,7 +7,7 @@ export const taskRouter = createProtectedRouter()
     input: z.object({
       id: z.string(),
       name: z.string(),
-      createdAt: z.date()
+      createdAt: z.date(),
     }),
     resolve: async ({ ctx, input }) => {
       const { id, name, createdAt } = input;
@@ -15,7 +15,8 @@ export const taskRouter = createProtectedRouter()
         data: {
           id,
           name,
-          createdAt
+          createdAt,
+          userId: ctx.session.user.id,
         },
       });
     },
@@ -24,7 +25,11 @@ export const taskRouter = createProtectedRouter()
   // READ
   .query('get-all', {
     resolve: async ({ ctx }) => {
-      return await ctx.prisma.task.findMany();
+      return await ctx.prisma.task.findMany({
+        where: {
+          userId: ctx.session.user.id,
+        },
+      });
     },
   })
 
@@ -36,6 +41,10 @@ export const taskRouter = createProtectedRouter()
     }),
     resolve: async ({ ctx, input }) => {
       const { id, isDone } = input;
+      const tastToUpdate = await ctx.prisma.task.findFirstOrThrow({ where: { id } });
+      if (tastToUpdate.userId !== ctx.session.user.id) {
+        return;
+      }
       return await ctx.prisma.task.update({
         where: { id },
         data: { isDone },
@@ -50,6 +59,10 @@ export const taskRouter = createProtectedRouter()
     }),
     resolve: async ({ ctx, input }) => {
       const { id } = input;
+      const tastToUpdate = await ctx.prisma.task.findFirstOrThrow({ where: { id } });
+      if (tastToUpdate.userId !== ctx.session.user.id) {
+        return;
+      }
       return await ctx.prisma.task.delete({
         where: { id },
       });
